app/requirements.txt
Django~=4.2
djangorestframework~=3.14.0
drf_yasg~=1.21.4
psycopg2-binary~=2.9.6
djangorestframework-simplejwt~=5.3.1
requests
pyxirr
elasticsearch~=8.14.0
black~=23.7.0